# Reduce memory allocation by 80%