@jmkeil 2023-05-04
Nice. Thanks for the PR.

Please let me know, if you want to be [mentioned in pom.xml](https://github.com/fusion-jena/JaroWinklerSimilarity/blob/03266936fa5350724979d78fef4213a20189ed0a/pom.xml#L40) in another way.