@uk8422 2023-05-04
> Please let me know, if you want to be mentioned in pom.xml

No, that is fine.

Thanks for the library! :)